refactor(suite): extract shared connect-permission helpers
Move getCoinLabel, the permission preview limit and the permission→icon map
into @suite-common/connect-popup so the web and native permissions UIs share
them. The web view now imports these instead of its local copies, and the
duplicated getCoinLabel util is removed.
