+++ suite-common/connect-popup/src/permissionsGrouping.ts
@@ -1,3 +1,4 @@
+import { networks } from '@suite-common/wallet-config';
 import { type PermissionRequest } from '@trezor/connect';
 import { unique } from '@trezor/utils/src/unique';
 
@@ -7,6 +8,41 @@ export type GroupedPermissions = {
     coin?: string;
     permissions: PermissionRequest['permission'][];
 };
+
+// When a coin group is collapsed, at most this many permission icons are
+// previewed next to its heading (shared by the web and native UIs).
+export const PERMISSION_PREVIEW_LIMIT = 6;
+
+/**
+ * Map a `PermissionRequest.coin` (which is `coinInfo.shortcut`, e.g. `BTC`,
+ * `ETH`, `LTC`) to a human-readable network name. Falls back to the upper-case
+ * shortcut when the coin is not known to suite (e.g. an altcoin recognised
+ * only by `@trezor/connect`).
+ */
+export const getCoinLabel = (shortcut: string): string => {
+    const key = shortcut.toLowerCase() as keyof typeof networks;
+
+    return networks[key]?.name ?? shortcut.toUpperCase();
+};
+
+// Icon shown for each permission. These names exist in both the web icon set and
+// the mobile icon font, so the web and native permission UIs render identical
+// glyphs. Typed as plain strings here (the shared package has no icon-library
+// dependency); each platform validates the value against its own `IconName` at
+// the call site.
+export const permissionIcons = {
+    read_address: 'eye',
+    read_xpub: 'key',
+    read_account_info: 'wallet',
+    read_settings: 'slidersHorizontal',
+    read_features: 'cpu',
+    sign: 'signature',
+    sign_message: 'notePencil',
+    verify_message: 'sealCheck',
+    management: 'gearSix',
+    push_tx: 'broadcast',
+    internal: 'cube',
+} as const satisfies Record<PermissionRequest['permission'], string>;
 
 export const permissionsAreCovered = (
     requested: PermissionRequest[],
